refactor: share the scheduler's solver-agnostic input handling
The two backends build the same model in two representations, so the model
construction is necessarily written twice. Everything around it was too:
199 lines were byte-identical between linear_optimization.py and
highspy_optimization.py -- argument normalisation, stock groups and their
validation, legacy commitment conversion, the sub-commitment split,
device_group_lookup, the convex-curve check, the Big-Ms, band validation,
the HiGHS option profile, and the cost/schedule assembly. None of it has a
solver in it, and keeping it twice meant the two paths could drift apart on
input handling, which the equivalence tests are not aimed at.

Move it to a new scheduling_problem module: prepare_scheduling_problem()
returns a SchedulingProblem that both backends unpack, plus solver_options()
and the result-assembly helpers. Duplication between the backends drops from
199 to 40 lines, and those 40 are the shared signature and the call itself.

Two deliberate changes while moving:

- commodity_devices becomes a cached_property. It is a per-row scan that
  only the Pyomo ems_flow_commitment_equalities needs, so computing it
  eagerly would put a real cost on the fast path. Pyomo pays what it did
  before; the direct backend pays nothing until it needs it (see #2355).
- initial_stock_of() casts its index to int, as the direct backend already
  did. The Pyomo version raised TypeError on the numpy float device indices
  a commitment's "device" column can carry.

The empty-commitments case also stops raising on pd.concat([]), which
previously made the Pyomo path crash where the direct path coped.

Verified: 265 passed, 3 xfailed across the planning suite under all three
solver parameters, plus the scheduling-job and API schedule tests.
